refactor(translator): group core domain modules under core/domain
Move the payload-free translation-domain modules — content-projection, field-config,
provenance, translation-providers — into core/domain/ (tier 2). Pure path moves;
importer specifiers updated; exported symbols and public API unchanged.
