+++ packages/payload-plugin-translator/src/core/domain/content-projection/fingerprinter.test.ts
