+++ packages/payload-plugin-translator/src/core/domain/content-projection/idPath.ts
