+++ packages/payload-plugin-translator/src/core/domain/content-projection/translatableLeaf.ts
@@ -9,9 +9,9 @@
  */
 
 import { isFieldExcludedFromTranslation } from "../field-config";
-import type { FieldLike, LeafFieldLike } from "../kernel/field-traversal/types";
-import { collectSerializedLexicalTextNodes } from "../kernel/lexical/collectTextNodes";
-import { isSerializedLexicalRoot } from "../kernel/lexical/guards";
+import type { FieldLike, LeafFieldLike } from "../../kernel/field-traversal/types";
+import { collectSerializedLexicalTextNodes } from "../../kernel/lexical/collectTextNodes";
+import { isSerializedLexicalRoot } from "../../kernel/lexical/guards";
 
 /** The field types this package translates (mirrors `isTranslatableField`, payload-free). */
 const isTranslatableFieldType = (field: { type: string }): boolean =>
