+++ packages/payload-plugin-translator/src/core/domain/field-config/getFieldConfig.test.ts
