+++ packages/payload-plugin-translator/src/core/domain/field-config/index.ts
