+++ packages/payload-plugin-translator/src/core/domain/provenance/ProvenanceStore.interface.ts
