+++ packages/payload-plugin-translator/src/core/domain/provenance/index.ts
