+++ packages/payload-plugin-translator/src/core/translation-pipeline/stages/field-collector/driftGuard.test.ts
@@ -1,7 +1,7 @@
 import type { Field } from "payload";
 import { describe, expect, it } from "vitest";
 
-import { projectTranslatableContent } from "../../../content-projection/contentProjector";
+import { projectTranslatableContent } from "../../../domain/content-projection/contentProjector";
 import type { FieldLike } from "../../../kernel/field-traversal/types";
 import { OverwriteStrategy } from "../../strategies";
 import type { FieldChunk } from "../../types";
