+++ packages/payload-plugin-translator/src/server/modules/index.ts
@@ -1,2 +1,2 @@
-export * from "../../core/translation-providers";
+export * from "../../core/domain/translation-providers";
 export * from "../../core/translation-pipeline";
