+++ packages/payload-plugin-translator/src/translation-providers/openai/OpenAITranslation.provider.ts
@@ -4,7 +4,7 @@ import type {
   TranslationProvider,
   TranslationInput,
   TranslationOutput,
-} from "../../core/translation-providers/TranslationProvider.interface";
+} from "../../core/domain/translation-providers/TranslationProvider.interface";
 import { isObject } from "../../core/kernel/utils/isObject";
 import type { ChatModel } from "openai/resources/index.mjs";
